JsonFileDeserializer Refactor
